+++ .gitignore
@@ -31,6 +31,7 @@ report.toc
 
 test.ipynb
 plotting.ipynb
+plots.ipynb
 new.ipynb
 
 __pycache__
